2.1.0.8

+++ src/rpc/rpcnoirnode.cpp
@@ -633,7 +633,7 @@ UniValue listnoirnodes(const UniValue& params, bool fHelp) {
             obj.push_back(Pair("version", mn->nProtocolVersion));
             obj.push_back(Pair("lastseen", (int64_t)mn->lastPing.sigTime));
             obj.push_back(Pair("activetime", (int64_t)(mn->lastPing.sigTime - mn->sigTime)));
-            obj.push_back(Pair("lastpaid", (int64_t)mn->GetLastPaidTime()));
+            obj.push_back(Pair("lastpaid", mn->GetLastPaidTime()));
 
             ret.push_back(obj);
         }
